+++ test/tracks/GenotypeMatrixTest.java
@@ -5,6 +5,7 @@ import static org.junit.Assert.*;
 import java.io.File;
 import java.io.IOException;
 import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
@@ -12,12 +13,14 @@ import org.junit.Before;
 import org.junit.Test;
 
 import coloring.Config;
-import coloring.ConfigKey;
 import coloring.Xterm256;
 import exceptions.InvalidColourException;
 import exceptions.InvalidConfigException;
 import exceptions.InvalidGenomicCoordsException;
 import exceptions.InvalidRecordException;
+import htsjdk.variant.variantcontext.VariantContext;
+import htsjdk.variant.vcf.VCFFileReader;
+import htsjdk.variant.vcf.VCFHeader;
 import samTextViewer.GenomicCoords;
 
 public class GenotypeMatrixTest {
@@ -28,32 +31,85 @@ public class GenotypeMatrixTest {
 		new Xterm256();
 	}
 
-	@Test
-	public void bigData()throws InvalidGenomicCoordsException, IOException, ClassNotFoundException, InvalidRecordException, SQLException, InvalidColourException {
+//	@Test
+//	public void bigData()throws InvalidGenomicCoordsException, IOException, ClassNotFoundException, InvalidRecordException, SQLException, InvalidColourException {
+//
+//		GenomicCoords gc= new GenomicCoords("1:572807-755079", 200, null, null);
+//		TrackIntervalFeature vcf= new TrackIntervalFeature("/Users/db291g/Downloads/ALL.wgs.mergedSV.v8.20130502.svs.genotypes.vcf.gz", gc);
+//		vcf.setNoFormat(false);
+//		long t0= System.currentTimeMillis();
+//		System.err.println(vcf.printToScreen().length());
+//		long t1= System.currentTimeMillis();
+//		System.err.println(t1-t0);
+//	}	
 
-		GenomicCoords gc= new GenomicCoords("1:572807-755079", 200, null, null);
-		TrackIntervalFeature vcf= new TrackIntervalFeature("/Users/db291g/Downloads/ALL.wgs.mergedSV.v8.20130502.svs.genotypes.vcf.gz", gc);
-		vcf.setNoFormat(false);
-		long t0= System.currentTimeMillis();
-		System.err.println(vcf.printToScreen().length());
-		long t1= System.currentTimeMillis();
-		System.err.println(t1-t0);
-	}	
-	
 	@Test
-	public void canInitMatrix() throws InvalidGenomicCoordsException, IOException, ClassNotFoundException, InvalidRecordException, SQLException, InvalidColourException {
+	public void canFilterWithJavascript() throws InvalidGenomicCoordsException, IOException, ClassNotFoundException, InvalidRecordException, SQLException, InvalidColourException {
 		
-		GenomicCoords gc= new GenomicCoords("1:572807-755079", 80, null, null);
-		TrackIntervalFeature vcf= new TrackIntervalFeature("test_data/ALL.wgs.mergedSV.v8.20130502.svs.genotypes.vcf.gz", gc);
+		VCFFileReader reader = new VCFFileReader(new File("test_data/info_formats.vcf.gz"));
+		VCFHeader vcfHeader = reader.getFileHeader();
+		reader.close();
+		
+		GenomicCoords gc= new GenomicCoords("1:17822074-17822184", 80, null, null);
+		TrackIntervalFeature vcf= new TrackIntervalFeature("test_data/info_formats.vcf.gz", gc);
+
 		List<IntervalFeature> linf = vcf.getIntervalFeatureList();
 
 		GenotypeMatrix gm= new GenotypeMatrix();
 		
-		// No data at all
-		assertTrue(gm.printToScreen(true).isEmpty());
+		// No filter
+		assertEquals(2, gm.printToScreen(true, linf, 80, vcfHeader).split("\n").length);
+
+		// 
+		gm.setJsScriptFilter("{DP} >= 100");
+		String x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertEquals(1, x.split("\n").length);
 		
-		gm.makeMatrix(linf, 80);
-		String x= gm.printToScreen(true);
+		// Filter by ALT: Note array slicing
+		// NB: Although all samples satisfy ALT, only sample2 also satisfies DP
+		gm.setJsScriptFilter("{ALT}[0] == 'G' && {DP} > 5");
+		x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertEquals(1, x.split("\n").length);
+
+		// JS script must return boolean
+		boolean pass= false;
+		try{
+			gm.setJsScriptFilter("10 + 3");
+			gm.printToScreen(true, linf, 80, vcfHeader);
+		} catch(InvalidGenomicCoordsException e){
+			pass= true;
+		}
+		assertTrue(pass);
+
+		GenotypeMatrix gm2= new GenotypeMatrix();
+		int i= 0;
+		while(i < 1000){
+			gm2.setJsScriptFilter("{ALT}[0] == 'G' && {DP} > 5");
+			String y= gm2.printToScreen(true, linf, 80, vcfHeader);
+			System.err.println(i);
+			i++;
+		}
+	}
+	
+	@Test
+	public void canInitMatrix() throws InvalidGenomicCoordsException, IOException, ClassNotFoundException, InvalidRecordException, SQLException, InvalidColourException {
+		
+		VCFFileReader reader = new VCFFileReader(new File("test_data/ALL.wgs.mergedSV.v8.20130502.svs.genotypes.vcf.gz"));
+		VCFHeader vcfHeader = reader.getFileHeader();
+		reader.close();
+		
+		GenomicCoords gc= new GenomicCoords("1:572807-755079", 80, null, null);
+		TrackIntervalFeature vcf= new TrackIntervalFeature("test_data/ALL.wgs.mergedSV.v8.20130502.svs.genotypes.vcf.gz", gc);
+
+		List<IntervalFeature> linf = vcf.getIntervalFeatureList();
+
+		GenotypeMatrix gm= new GenotypeMatrix();
+		
+		// No feature at all
+		assertTrue(gm.printToScreen(true, new ArrayList<IntervalFeature>(), 80, vcfHeader).isEmpty());
+		
+		// gm.makeMatrix(linf, 80, null);
+		String x= gm.printToScreen(true, linf, 80, vcfHeader);
 		
 		// Check sample name
 		assertTrue(x.startsWith("HG00096"));
@@ -77,16 +133,17 @@ public class GenotypeMatrixTest {
 		List<IntervalFeature> linf = vcf.getIntervalFeatureList();
 
 		GenotypeMatrix gm= new GenotypeMatrix();
-		gm.makeMatrix(linf, 80);
 		gm.setSelectSampleRegex("96|99");
+		// gm.makeMatrix(linf, 80, null);
 
-		String x= gm.printToScreen(true);
+		String x= gm.printToScreen(true, linf, 80, null);
 		String[] rows = x.split("\n");
 		assertEquals(2, rows.length);
 		
 		// Exclude all samples
 		gm.setSelectSampleRegex("^$");
-		assertTrue(gm.printToScreen(true).isEmpty());
+		// gm.makeMatrix(linf, 80, null);
+		assertTrue(gm.printToScreen(true, linf, 80, null).isEmpty());
 	}
 	
 	@Test
@@ -106,8 +163,8 @@ public class GenotypeMatrixTest {
 		List<IntervalFeature> linf = vcf.getIntervalFeatureList();
 
 		GenotypeMatrix gm= new GenotypeMatrix();
-		gm.makeMatrix(linf, 80);
-		String x= gm.printToScreen(true);
+		// gm.makeMatrix(linf, 80, null);
+		String x= gm.printToScreen(true, linf, 80, null);
 		assertTrue(x.isEmpty());
 	}
 
@@ -119,8 +176,8 @@ public class GenotypeMatrixTest {
 		assertTrue(linf.size() > 0); // Make sure we do have some features otherwise the test is meaningless.
 		
 		GenotypeMatrix gm= new GenotypeMatrix();
-		gm.makeMatrix(linf, 80);
-		String x= gm.printToScreen(true);
+		// gm.makeMatrix(linf, 80, null);
+		String x= gm.printToScreen(true, linf, 80, null);
 		assertTrue(x.isEmpty());
 	}