+++ test/tracks/TrackIntervalFeatureTest.java
@@ -37,13 +37,13 @@ public class TrackIntervalFeatureTest {
 		TrackIntervalFeature tif= new TrackIntervalFeature("test_data/hg19_genes_head.gtf", gc);
 		tif.printToScreen(); // This is to populate the ideograms.
 		
-		Map<String, String> colorForRegex= new HashMap<String, String>();
-		colorForRegex.put("DDX11L1", "216");
+		Map<String, Argument> colorForRegex= new HashMap<String, Argument>();
+		colorForRegex.put("DDX11L1", new Argument("216", false));
 		tif.setColorForRegex(colorForRegex);
 		assertTrue(tif.printToScreen().contains("216"));
 
 		colorForRegex.clear();
-		colorForRegex.put("WASH7P", "233"); // 233:grey7 (almost black)
+		colorForRegex.put("WASH7P", new Argument("233", false)); // 233:grey7 (almost black)
 		tif.setColorForRegex(colorForRegex);
 		assertTrue(tif.printToScreen().contains("233"));
 		assertTrue(tif.printToScreen().contains("216"));